Fail when trying to ensure an installed version on FreeBSD
The two package providers for FreeBSD are not able to ensure that a
specific version of a package is installed because of how ports/pkg_add
work. They can only ensure the version from the specified FreeBSD
release is installed or absent.

For more information, see:

http://docs.puppetlabs.com/references/stable/type.html#features-4

To make the class easier to understand, let's not do something that
users don't expect when they're requesting installation of a specific
version. We'll explicitly fail and tell users why so that they can
adjust how they use the class/module.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>