fix default value case for $cron_time
with the current code it's impossible to use the automatically generated
cron time spec since to get there you need to set $cron_time to false,
and this one in turn gets converted to an integer. This throws an error.

To fix this, we'll base the second timing to be 30mins after the first
one.